+++ Core/Constants.cs
@@ -4,21 +4,24 @@
     {
         #region Public Fields
 
-        public const double DividerThreshold = 100;
+        public const int ClipPositionFactor = 50;
 
+        public const int DelayMin = 80;
+
+        public const int DurationMax = 1000;
         public const int DurationUpdateTitle = 500;
-
-        public const int MaxDuration = 1000;
-        public const int MaxQueueSize = 20;
-        public const int MaxThreshold = 100;
-
-        public const int MinDelay = 80;
-        public const int MinQueueSize = 1;
 
         public const int PortHttpWebServer = 5003;
         public const int PortHttpWebSocket = 9000;
 
+        public const int QueueSizeMax = 20;
+        public const int QueueSizeMin = 1;
+
         public const int RecognitionMilliSecondsMax = 1000 * 60 * 10;
+
+        public const double SelectionFontSize = 16.0;
+        public const double SelectionThicknessActive = 3.0;
+        public const double SelectionThicknessNormal = 1.0;
 
         public const string SettingsFileNameDefault = "userSettings.json";
 
@@ -28,6 +31,12 @@
         public const string TabBoard = "Board";
         public const string TabClips = "Clips";
         public const string TabSamples = "Samples";
+
+        public const double ThresholdDivider = 100;
+        public const int ThresholdMax = 100;
+
+        public const double ZoomMax = 5;
+        public const double ZoomMin = 0.5;
 
         #endregion Public Fields
     }