+++ Core/Events/Clip/ClipsOrderedEvent.cs
@@ -1,0 +1,8 @@
+﻿using Prism.Events;
+
+namespace Score2Stream.Core.Events.Clip
+{
+    public class ClipsOrderedEvent
+        : PubSubEvent
+    { }
+}