+++ Core/Events/Video/VideoCenteredEvent.cs
@@ -1,0 +1,8 @@
+﻿using Prism.Events;
+
+namespace Score2Stream.Core.Events.Video
+{
+    public class VideoCenteredEvent
+        : PubSubEvent
+    { }
+}