+++ Modules/VideoModule/Views/SelectionView.axaml.cs
@@ -1,13 +1,15 @@
-using Avalonia;
 using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
 
 namespace Score2Stream.VideoModule.Views;
 
 public partial class SelectionView : UserControl
 {
+    #region Public Constructors
+
     public SelectionView()
     {
         InitializeComponent();
     }
+
+    #endregion Public Constructors
 }