+++ Additionals/ChangeLog.md
@@ -7,6 +7,7 @@
 * Noise removal added (https://github.com/budul100/Score2Stream/issues/9)
 * Zoom and pan of video view implemented
 * Templates are independent of samples
+* Samples can be ordered
 * Request is shown if a sample with value shall be removed
 * Number of periods is saved in settings
 * Menu entries reordered